Create Movie Ticket Booking.py
5.	Write a Python program to implement the following Scenario:
Create a class movie with name of the movie and movie available in the theatre [“PVR”,” INOX”,” CARNIVAL”] based on the list provided user will select anyone theatre for seeing movie. After that provide the number of tickets available for a movie and type (gold:45 Recliner:20 and Silver:40). Now from this user will select any one type of ticket and ask how many numbers of tickets user would like to book of that type. After that ask user whether he wants to add meal yes or no. After this calculate the total price of the ticket based on number of tickets+meal (if yes) +tax (18%). Display the proper information of movie. Do the processing for 2 users.